Fix or silence compiler warnings about unreachable code, narrowing conversions, and signed/unsigned comparisons
(cherry picked from commit 39a708bb0a1aa62a7a72e0059bdf5e0eae6fb8bb)

+++ example/grep/grep.cpp
@@ -36,7 +36,7 @@ boost::regex_constants::syntax_option_type flags = boost::regex_constants::basic
 boost::regex re;
 boost::smatch what;
 std::string current_file;
-int file_count;
+size_t file_count;
 
 void process_stream(std::istream& is)
 {
